users_microservice/app/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-filter==23.2
Markdown==3.4.4
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-dotenv==1.0.0
django-cors-headers==4.2.0
requests==2.31.0
Pillow==10.0.0
django-storages==1.14
django-resized==1.0.2
pytils==0.4.1
django-extensions==3.2.3


